add sentiment distribution chart (in-line)
centre chart titles
