+++ dashboard/csv/overview.csv
@@ -1,0 +1,66 @@
+﻿unrecognised_msgs
+"[""I have a question about my recent transactions."",  ""How can I set up e-statements?"",  ""Can you help me with my credit card rewards?""]"
+"[""I'm locked out of my online banking account."",  ""What's the process for applying for a credit card?"",  ""I'd like to update my contact information.""] "
+"[""I suspect fraud on my account. What should I do?"",  ""Can you assist with changing my contact information?"",  ""I'd like to request a credit limit increase.""]"
+"[""I received a suspicious email. Is it from the bank?"",  ""How can I enroll in e-statements?""] "
+"[""What's the bank's holiday schedule?"",  ""I'd like to change my online banking password.""] "
+"[ ""Is there a branch nearby I can visit?"",  ""I lost my checkbook. Is there a branch nearby I can visit?""] "
+"[""Can you provide information on investment options?""] "
+"[""I received a credit card statement. Can you clarify the charges?""]"
+"[""I suspect fraud on my account. What should I do?"",  ""Can you assist with changing my contact information?"",  ""I'd like to request a credit limit increase.""]"
+"[""I have a question about a recent transaction."",  ""I'd like to sign up for mobile banking.""]"
+"[""I'm interested in opening a joint account. Is there a branch nearby I can visit?""] "
+"[""I received a suspicious call claiming to be from the bank. Is it legitimate?"",  ""How can I request account statements?""]"
+"[""I'm having trouble with mobile check deposit.""] "
+"[""I'm experiencing issues with online banking."",  ""What are the current exchange rates?""] "
+"[""How can I open a new account?"",  ""What documents are required for account opening?""]"
+"[""I'd like to change my account PIN."",  ""How can I update my contact information?""]"
+"[""I lost my debit card. What should I do?"",  ""Can you block my card?""]"
+"[""Can you send me my account statement?"",  ""How can I access e-statements?""]"
+"[""I've moved. How do I update my address?"", ""What's the process for changing my address?""]"
+"[""I'd like to apply for a credit card."",  ""What are the credit card options available?""]"
+"[""I see an incorrect charge on my statement."",  ""How can I dispute a transaction?""]"
+"[""How can I open a joint account?"",  ""What are the joint account requirements?""]"
+"[""Tell me about personal loan options."",  ""What's the loan application process?""]"
+"[""I'd like to apply for a small business loan."",  ""What's the process for small business loan applications?""]"
+"[""Can I increase my credit card limit?"",  ""What's the criteria for credit limit increases?""]"
+"[""I want to set up a recurring transfer."", What's the daily transfer limit for wire transfers?""]"
+"[""Can you provide information on personal loans?""] "
+"I suspect fraud on my account. What should I do?,  ,  ""What's the process for closing an account?""] "
+"[""How can I set up automatic transfers?"", ""I need help with a stop payment request.""]"
+"[""I want to open a custodial account for my child."",  ""Can you provide information on credit counseling services?"",  ""I'd like to dispute a fee on my account.""]"
+"[""I'm having trouble with mobile check deposit."",  ""How can I request a new debit card PIN?"",  ""I need assistance with redeeming rewards points.""] "
+"[""I received a suspicious text message. Is it from the bank?""] "
+"[""I need help with a wire transfer."",  ""Can you assist with a loan application?"",  ""How can I open a new account?""]"
+"[""I'm having trouble with my debit card."",  ""Is there a branch nearby I can visit?"",  ""I lost my checkbook. What should I do?""] "
+"[""I want to open a custodial account for my child."",  ""Can you provide information on retirement accounts?""]"
+"[""What's the credit limit on the credit cards?"",  ""Are there any promotional offers for credit card applications?""]"
+"[""What types of accounts can I open?"",  ""Do I need an initial deposit to open an account?""]"
+"[""How much does it cost to stop a payment?"",  ""What's the process for stopping a check payment?""]"
+"[""What's the minimum credit score for a personal loan?"",  ""Are there any prepayment penalties for personal loans?""]"
+"[""How can I reset my mobile banking app password?"",  ""Is the mobile app available for both Android and iOS?""]"
+"[""What's the loan-to-value ratio for home equity loans?"",  ""Are there any closing costs associated with home equity loans?""]"
+"[""How is the interest rate for loans determined?"",  ""What's the maximum loan term available?""]"
+"[""What's the minimum balance requirement for new accounts?"",  ""Can I open a joint account with a family member?""]"
+"[""How can I change my PIN online?"",  ""Is there a fee for changing the PIN?""]"
+"[""What should I do if my card is lost or stolen?"",  ""Can I expedite the replacement of my lost card?""]"
+"[""How often are account statements generated?"",  ""Can I receive electronic statements instead of paper ones?""]"
+"[""What documentation is needed to update my address?"",  ""Is there a fee for changing my address with the bank?""]"
+"[""What investment products does the bank offer?"",  ""What's the minimum investment amount for brokerage accounts?""]"
+"[""What's the cutoff time for same-day wire transfers?"",  ""Are international wire transfers supported?""]"
+"[""Can I open an account online?"",  ""What account types are suitable for savings?""]"
+"[""How can I reset my PIN if I forget it?"",  ""Is there a limit to the number of PIN changes I can make?""]"
+"[""Can you help me with my credit card rewards?"",  ""I have a question about my recent transactions."",  ""How can I set up e-statements?""]"
+"[""I need assistance with a foreign currency exchange.""]"
+"[""I'd like to request a credit limit increase."",  ""Can you assist with changing my contact information?"",  ""I suspect fraud on my account. What should I do?""]"
+"[""I need to initiate a wire transfer."",  ""What information do I need for the transfer?""]"
+"[""I'd like to apply for a credit card."",  ""What credit card options are available?""]"
+"[""How can I open a new account?"",  ""What documents are required for account opening?""]"
+"[""I need to stop a payment on my account."",  ""How can I request a stop payment?""]"
+"[""I'm having trouble with the mobile banking app."",  ""Can you help me troubleshoot the issue?""]"
+"[""What retirement account options are available?""]"
+"[""Tell me about investment options."",  ""Can I schedule a meeting with a financial advisor?""]"
+"[""I'd like to change my online banking password."",  ""How do I reset my password?""]"
+"[""How long does it take to get approved for a credit card?"",  ""What's the credit limit on the cards?""]"
+"[""How often should I change my online banking password?"",  ""What are the password requirements for online banking?""]"
+"[""What's the maximum loan amount for small business loans?"",  ""What's the interest rate for small business loans?""]"
